Codesniffer updates
